fixed look on diff sizes, results layout
The quiz intro and initial results layout looked bad on small phones.
Large devices also were a hassle for results. Set up the textview in
quiz intro to scroll because it's long. Set up the text and score to
resize heights and be ok with more room via a space object in xml.